refactor: relocate classified-call-site types into mount_graph, delete call_site_classifier
Closes #77. CallSiteType/ClassifiedCallSite/HandlerArgument were the only
survivors of the long-deleted LLM call-site classifier, kept alive solely
because mount_graph imported them. They now live with their one consumer;
src/call_site_classifier.rs is gone, along with the lib.rs/main.rs module
declarations. (The prompt-leak baseline no longer carried an entry for the
file, so no ratchet change.)

Also fold the clippy manual_contains fix into the #102 regression test.
