+++ src/lib.rs
@@ -2,7 +2,6 @@ pub mod agent_service;
 pub mod agents;
 pub mod analyzer;
 pub mod app_context;
-pub mod call_site_classifier;
 pub mod call_site_extractor;
 pub mod cloud_storage;
 pub mod config;
