+++ src/main.rs
@@ -2,7 +2,6 @@ mod agent_service;
 mod agents;
 mod analyzer;
 mod app_context;
-mod call_site_classifier;
 mod call_site_extractor;
 mod cloud_storage;
 mod config;
